requirements.txt
torch==1.7.1
torchvision==0.8.2
torchaudio==0.7.2
black==19.10b0
pandas==1.2.1
scikit-learn==0.23.2
matplotlib==3.3.2
seaborn==0.11.1
tensorboard==2.3.0
pyyaml==5.4.1
tqdm==4.59.0